Cleaned warnings and added compilation and usage hints.

+++ WirelessSignalAdjuster.cpp
@@ -99,7 +99,7 @@ void WirelessSignalAdjuster::getSignalParameters() {
         defaultTXPower = getTXLevel();
         std::cout << "Default TX power: " << defaultTXPower << " [dBm]\n";
     }
-    for (int i = 0; i < maxTXPowerNumber; i++) {
+    for (size_t i = 0; i < maxTXPowerNumber; i++) {
         std::cout << "Channel [" << i + 1 << "] " << wirelessInterfaceParameters.freq[i].e - 0x100 << " dBm\n";
     }
 }
